Updated setup instructions
I updated the setup instructions after going through them to help
with any blockers and potentially make things a little easier.

- I added instructions to install rvm (although it might be
ok to assume it's installed)
- I added instructions to install postgres on OSX
- I changed the yarn engine version to anything over 1.3.0, because
the newest version is 1.5.0 so will not run if it requires 1.3.x.
- I added an nvmrc file, which is similar to an rvmrc file and allows
you to use type `nvm use` to get to the right nvm version (and can
let you setup a terminal to automatically go to the correct nvm
version).

e Please enter the commit message for your changes. Lines starting