Include resource in roles API fetch
So when you go to the RolesMenu, you fetch the roles based on the
collection ID. But if from there, you go to another page, you'll
fetch roles for another collection, and now you will have both on
the front end. So this means I want to filter the roles on the front
end by collection ID by including the resource.

It might be a little faster to just include the resource id rather
then the whole resource.