Fixed some roles edge cases
Related to Role creation in different states. Before I was modifying
the data on the front end but this is really difficult because of the
many different states from the API. This works from a data perspective
but the UI has some messy states. We might want to set a loading state
or modify the store in a way so the data isn't changed locally until
the final fetch.