Smarter alert bar functionality
When a notification comes in with a created date before 30 seconds
ago, it will add attempt to add it to a recent notifications array in
the api store. If this isn't read yet, it will be shown to the user in
the comments section. After 3 seconds, it will be removed from recent
notifications.