Fix id vs meta.id trouble
id doesn't work as it should.  We got everything working as it should
be removing id from the attributes of the serializers (which were
causing a setter / getter to be set on the models) so that we could
add our own id getter on `BaseRecord` which reaches in to meta to get
the id in a conssistent way.