Account page larger than maxwidth in calc on drag
Since we're positioning the card ourselves now, we have to position
it within our own max width and margins, which are 1320. Otherwise
on a large screen, larger then 1320, when you drag the card it will be
moved way off to the right because the value is using the whole page
width when the card coordinates are within the max width margins.