Fix menu positioning on small screens
Switched the menu positioning to use clientX as opposed to screenX.

ScreenY is still required for the vertical positioning because it takes
scrolling position into account. The clientX provides a consistent
number for the grid cards within different pages sizes, allowing the
horizontal positioning to work.

Additionally, I fixed another bug where right clicking a card would pop
it above the page header. This does mean that you can't click the page
header to get out of the context menu, I'm not sure which part is worse?