Image actions should show up on templates
The show as cover action should be allowed for anybody that is editor
of the collection, since that's essentially what your editing when
you pick a cover image. The image contain should be for content_editors
of cards, so you can edit it when you're using a templte, but not when
you only have view access to the card.