Attempt tofix org creation
Josh and I theorized that in certain circumstances on production, the
database commit was happening slower, leading to the `create_groups`
function to be called and saved in a timing where the slug isn't checked
correctly. Changing when the `create_groups` method happens may
fix this issue.

Changing to different way of same previous problem

Fix tests for org slug problem