Fix for IE
A vendor library was including code that breaks in IE. It used a
`for in` loop with a `const` definition. While IE can handle `const`
it can't handle it in a `for in` loop.

The fix was to quickly move this library internally to our code so
it gets built.