Fix slow hovering over blank spots.
This was a strange case. Previously when hovering over the blank spots,
it would only render the blank spot blue card if you stopped moving your
mouse for a milisecond or so. I think this is relating to it having
to render so quickly as you move your mouse around that it couldn't
keep up.

With this throttle it works perfectly and is better for performance.