Fix/Add overdue and regular banner unit tests
Had to use a weird pattern for testing render props:

```
const wrapper = render()
        const RightComponent = wrapper
          .find('StyledBanner')
          .prop('rightComponent')
        const html = shallow(RightComponent)
```

We shallow render the component we are testing using `render()`, and
then find the render prop of the component that returns HTML, which  we
then shallow render (like `Foo`, not `<Foo />`). Then we can operate on
the HTML from the render prop like we would on the larger component.

See here for more:
https://medium.com/@dferber90/test-a-render-prop-6a44e02f4c39