Fix confirm payment modal rendering logic
**What**
Revise logic to show correct text based on having payment method and/or
being an admin.

**Why**
Feedback editors were getting wrong messages when attempting to launch
their tests.

**Note:**
Bug introduced in this [ticket](https://trello.com/c/Umqk2s5w/1271-1-the-feedback-editor-must-agree-to-the-price-when-launching-a-test) as part of this [PR](https://github.com/ideo/shape/pull/551).