Adds support for new collection cover with text
This uses a text item in the collection (which could be hidden) and
set it to the current cover item. In then adds a new cover type to
deal with this special case of rendering both collection cover and
the text item cover within it. It applies some special styling to
the text item for this state.