Only create group if part of organization
We changed the way this worked so groups are only created when an
organization is passed along. The callback will happen again when
an organization_id is assiged to the ground on the network, so it
will enter this code path and create the group then.

This ensures no groups will be created on Shape with no organization.