Redirect logged out users: user settings => login
Add client-side redirect guard clause to user settings component.

Visitors were getting an "Oh no" (client side 500 error) message when
following a link from an email.

This was due to the component checking if a user is on a mailing list.
I went with a client side redirect since we don't currently have a
backend route for user_settings – it's only through React router.