Allow quill converter to take images
Which are required for the creative difference collection covers
which have an icon.

This follows the standard quill structure. It's not a feature that
would be available to users yet, so testing it requires manually
setting the `data_content` field in the database for an existing
text item. Here is a working example:

```
{"ops"=>
    [{"insert"=>"Hello cdelta"},
     {"insert"=>{"image"=>"https://ideo-sso.s3-us-west-2.amazonaws.com/headers/creative_difference_logo_small.png"}},
     {"insert"=>"\n", "attributes"=>{"header"=>1}},
     {"insert"=>"some other text\n"}],
   "version"=>23},

```