Allow application bots to move pinned cards
The only app bot, c delta, is creating lots of pinned items manually
and then has to move and sometimes delete them. It would be very
difficult for the bot to track whats pinned, un-pin it, and then re-pin
it, so we're just giving it the ability to move around pinned cards,
which should be safe because it's a robot and never makes mistakes.