Have collection covers dynamically set height
Which will be used for CDelta collection covers which use text
items. These text items can be of varying heights meaning we need
to dynimically set the height, otherwise the last part of the cover
could be cut off.

This should mainly affect CD covers because they use a text item
as the cover text.