Fix marketing pricing page based on data structure
The data structure for the pricing page was differently set up in
firebase than initially thought, so I modified the code to support
that structure.

It's very difficult and time-consuming to modify the structure in
firebase, so this code is a little less organized than desired to
accomodate that. Once we move to a better backend content management
solution, this can be improved.