Change use of identifier field on dataset
We need a translatable field for the UI for datasets to use on legend
items, which is different from the identifier, so now `name` is the
UI translatable field, and the previous `identifier` is an actual
code identifier for us in code. This needed to be done because
before, `identifier` was being used for both but broke the selecting
functionality when we added translations.