Only close open card menu if card not a text item
__Why:__
* Allow user to right-click on a different text selections

Adding this change provides a more natural interaction for the user by
allowing them to select some text and then, if they right-click on a
different piece of text, select the new piece of text without having to
open and close the text action menu with multiple right clicks.

We chose this implementation due to issues with the right click event
occurring before the on selection change event. In the future we may
revise our approach to use timestamps in order to handle this better.

The alternative was to have to right click twice for a subsequent text
selection, once to close the menu and once to re-open it. This felt
clunky, leading us to apply the fix in this commit.

__This change addresses the need by:__
* Keep card menu open if it is already open for a text item