WIP: connect comment button to notification log
The `Activity::handleClick` approach that was suggested seems difficult
to use since it requires multiple levels of prop drilling.

Is there a way to access the correct comment thread without having to
have access to the notifications like NotificationsContainer?

If not, perhaps we should reconsider how we are passing data around.