Allow legend cover to be scrollable
Because for CDelta we're changing the behavior so the legends
are 1x1, instead of 1x2. This means they might need to be scrollable
if you are adding a lot of comparisions.

This should rarely affect the existing 1x2 legends because you would
probably never add enough comparisons for scrolling to kick in.