Create consistent tooltip styles on area charts
They differed for a few reasons. One part, is the smaller chart style,
when the area is only 1, the corner radius value is too small relative
to the size the tooltip will be, so it has to be made larger. This
is the same problem with the font size, which had to be made even
larger to match the same font size as the larger report item charts.