Correct tooltip offsets for all charts
The offset ensures when chart points that are to the far left or
right of a chart, the tooltip will not appear outside of the chart
visible area, but will instead but nudged to the right or left. These
were working for certain charts but not others, so this code essentially
uses the chart width to get the right nudge values for different
chart sizes.