Ensure secondary CD datasets renders as line
This changes the logic so the secondary CDelta datasets render as
a line, even when their chart type is "area". This has to be done
because CDelta uses shared datasets for everything, including its
comparisons, so comparison datasets have `chart_type` 'area'. But
when they are used as a comparison vs a primary dataset, they should
change to a line type to visually display in the charts.