Apply default next card order if no cards
In the case that theres no cached last order and no collection cards,
this ensures the default last order is 0, which fixes the tests as
well as the case that this happens.

Also fixes some JS styling issues that was somehow not saved in the
last commit on this branch.