Add `style` to legend item
This is required for CDelta. CDelta uses shared datasets for all
of it's charts, which means for the first chart in the dashboard,
the chart for all the creative qualities, the primary dataset is
just a random quality. The selected comparison in the legend item
then uses the dataset's fill color, but that's not really correct
because this dataset really represents the whole BU. This change
allows the legend item to override the fill color of the comparisons
for just this use case.