Filtering UI for collection filters
This isn't hooked up to real actions yet, but the filters apprear
on collection pages. It makes some global changes to the pills that
we want to be consistent.

Introduces new models for collection filters and a user collection
filter for when you selected a filter in a collection on and off.