Send comment subject preview link to correct view
__Why:__
* Connect thumbnail preview of comment subject to original context in
which the subject was commented on, e.g., to the collection a user was
viewing when they commented on a given item/collection

__This change addresses the need by:__
* Differentiate between `threadRecord` and `subjectRecord`, where the
`threadRecord` acts as a source of truth for the context in which the
comment was made

__Notes:__
A comment will always belong to a comment thread, so we use the record
(item or collection) associated with the comment thread as the location
or context in which the comment was made. So if Person A comments on a
text item from the full page view and Person B comments on the same text
item from a collection view, their comment preview thumbnails will take
users to different locations when clicked.