More UI and MUI Chip refactor
This code removes the MUI `Chip` component from being used for our
own `Pill` UI, instead using our own componenent. This was done to
keep the pills in the tag modal / new filter modal consistent with
the pills elsewhere. The tag and filter modal use the ReactTags library
which allows you to pass in your own component for styling. If we passed
in the existing `Pill`, which was a styled `Chip`, we woudldn't be able
to customize it as much because ReactTags has a specific way it passes
through props to the component. Now that we use our own component we
can customize it for any use case, including the ReactTags one. It makes
it easier to maintain in the future as well.

Also adds the UI for the filter search modal.