Collection filters create and destroy
This uses a nested route under collection to create and destroy
the filters. This was done because the filters are always rendered
through the collection, so both methods will pass the whole collection
back to the front end to re-render it with the correct filters. The
authorization also happens through the collection as you have to have
edit access on the collection to do anything to the filters.

I decided to keep the filter modal a bit more generic, so it doesn't
operate with actual collection filters, but uses the more generic
tag data structure for everything (label, symbol, etc). This means
it could be reused for different type of filtering more easily and
the top-level CollectionFilter component handles any CollectionFilter
data manipulation.

Some of the routes will likely change as more of this gets figured
out.