User selectable filters
The UI calls for allowing users to (un)select the filters on a
collection and have those results saved, which will be stored in
its own `UserCollectionFilter` table.

The (un)selection is done with an API route nested under
`collection_filters` as they always belong to one. I set up two
separate select and unselect routes (vs a selected being a param)
because that's a pattern we've used elsewhere and makes it very clear.
The route always uses current user to find the record, which means
it doesn't need authorization, as theres no way to get to another
user's records. It does a find_or_create because the user filter
record get's created on the first try (an unselect), and then
keeps the user's preference saved from there.

The serializable collection filter keeps the selected state by looking
up if the current user has a user filter, and then gets the selected
state from that. This keeps the data models on the front end very
simple.