Unselect all filters button
This button unselects all filters to show all records in the collection.

I switched the logic for the toggleSelected method so when you pass
in false, it will unselect them all and vice versa. This makes it
easier to understand as you're sending in the action you want as
before, you were passing the current state and it was toggling it
from that. That would have meant that when you want to unselect them
all you would call the toggle method with `true` which is much less
clear to me, hence the switch.