Tests for more filtering components
I mainly tested branching logic in the render methods and event
handlers that have logic in them besides passing an event up. For
other simple handlers, there should be some sort of integration test
to ensure everyting is hookup up correctly as opposed to doing that
test in a unit test like this.