Messy, slow, first-pass at collection filtering
Get's filtering working by searching for Collection's and Items,
getting the ids and getting all CollectionCard's from there.

Because the correct filter selected state is already on the front end
as it's set in the collection_filter serializer, I decided to just
pass those filters back up to the backend with a query string. Otherwise
the backend would have to join the CollectionFilters to the user filters
itself to determine which ones are selected, which is duplicated work.

Could clean up:
- Getting the data back from searchkick, it's messy right now and is
not the straight Collection, Item objects
- Making the query faster? Less looping if possible
- Moving things to own methods, move things to shared places if we
use this pattern elsewhere (of getting all cards for collection/item
ids)