FIlter collection card endpoint by filter query
The filter query is assembled on the front end and passed on the
`collection_cards` endpoint.

The endpoint psased the query off to collection cards filter, which
will do a elastic search when it's present. This search uses an `or`
condition to search all records and linked records within the collection
to ensure it gets collection card links. If it just used the collections
parent id as previously done, it would not get the links as they
are technically linked to a different parent collection.