Adds collection card refreshing to front end UI
When any of the collection filters change, the UI should refetch
collection cards and show a loader so the user knows their action
of modifying the filters is doing something.

I implemented this with a shared method which wraps any of the updates
to the filter updates so it's all consistent.