Add autocomplete to filter search
The search term search will show how many results your search
brought up in the collection. The tag search will get all current
tags on the current collection to use as suggestions to ReactTags.

We decided to just pass all the suggestions to ReactTags as opposed
to searching for them with a `word_start` search because there will
never be that many tags on a single collection of direct children.
If we ever wanted to do a more global tag search, we'd have to add
a search controller method for tags and configure elastic search
to just search tags.