Fix test change default values
The first change changes the default selected values to false in the
serial collection filter. We want the user to have all filters
unselected when they first visit the page, and then they select things
on.

The second is a change in behavior that I'm going back on. When there
is no current user (logged out state) the render_collection method
should not pass a nil current user to it's serializable through
expose or tests break.