Front end rendering of new charts
Changes chart design in a few ways such as removing padding.

Adds logic so as you add more bars and groupings to charts, the widths
shrink and the labels truncate(with a tooltip) so the render nicely.

The refactoring was so we could have a shared tooltip component which
sets some styling and values for VictoryTooltip in one place.