Fix logic of bar widths in complex graphs
The width of the bar has to be set correctly so that they all fit
correctly with a miniumum amount of padding (16px).

I decided to pretty simply get the max area the bars could take
figure out the min amount of padding needed and use the total columns
with the total groupings to determine what the width should be. This
should work for most cases.

The labels below the graphs use this same value to determine their
own size, which isn't perfect because they actually have more room
sometimes, but generally works.

Padding has to be added to the left and right of the bar equal to the
full width of the bar with all groupings. This always has to be at least
one bar widths length.