Fix click handlers for carousel items
- Make sure you can't ever edit text items in carousels, by passing
an uneditable prop all the way to text cover.

- Ensure that clicking on carousel items routes to the collection.
This was done by adding a routing store request in the carousel
which then gets passed down to the covers below. This was best as no
component above it really had a handler for it.

- Disable grid card handle click for carousels, otherwise when you
click on the carousel controls, that click handler woudl be triggered,
because it's above the controls in the DOM hierarchy, so we had to
disable it at that level.