Add open responses to correct collection
The open responses for an alias should go in the alias collection's
open response collection, which you get to through the survey_response
in the idenfitier.

This also renames the Open response collections with an OpenResponse
identifier, as to not confuse them with the "All Response" collection
which uses the Response identifier.