Ensure open responses are updated
They weren't being updated before, because on creation we weren't
saving the open response item on the question answer, which it needs
to look it back up to update it. This wasn't being save because before,
it was in a before_save callback, but since that got changed, and it
get's called from the test results services, it needs an explicit
save call.