Fix layout issues on small screens
The feedback designer had layout issues with it's hot edges on all
small screens. This was due to how we set a top on the element, which
we set to the known height of the element on larger screens when
the question controls and preview are side by side. But on mobile
they stack on top of eachother so that `top` property wasn't accurate.

Because we don't know how tall the question controls will be, we can't
determine where to vertically position the hot edge. So I moved the
rendering of the hot edge into the question holder so it's all contained
there and we can set the top to the top of that container.