Add collection filter back to submissions collection
Not sure how it was removed.

I did find that it needed special styling for when the filter bar is
active. When in a submissions collection, it is always on the same
line with the sort menu, where in a normal collection, it moves from
above the bar to below it when the bar is active.

I also updated the mobile styles so it moves to the line below the
sort menu, as it's very difficult to find room for them both.