Experiment: Adds VisibilitySensor to MovableGridCard
- when not visible, GridCard (inner) is replaced with a blank gray div.
- seems like it improves performance 5-10%? hard to tell, otherwise it 
so far doesn't seem worth the visual annoyance of seeing gray cards as 
you scroll up/down. 
- the other problem is that we can't fully make the Rnd card disappear 
because it has no positioning for the visibility sensor to know where it 
is; so all we can do is hide the inside (e.g. rendered cover) so perhaps 
that is also why the improvement is minimal.