Allow items to have subtitle hidden
This is done through a new cached attribute which was the easiest
and quickest solution.

It uses a new getter on item for getting the subtitle, which keeps
it in sync with the collection.

It also fixes the UI for the collection cover editor by allowing you
to see the subtitle in the editnig UI even if it's hidden because
that makes more sense than returning a blank string for it.