Fix sort position for different cases
We need both the filter button and the sort menu to always appear
to the right, so I put them both in another flex container that gets
pushed to the right side with a margin-left: auto. Then I kept the
other elements that should appear to the left in the parent FlexGrow
container. This should account for all the cases.