Increase width of CollectionCoverTitle IconHolder
This change increases the IconHolder component in CollectionCoverTitle.
It is intended to make a fix for TemplateIcon being too small, and has
the following side effects:

- all icons in CollectionCoverTitle are slightly larger
- TemplateIcon looks correct

The alternative to this is to revise the viewbox for TemplateIcon, but
that needs to then be adjusted by location. The change in this PR seems
more localized in that it only affects icons in cover titles.

That said, if the increased icon size is a problem and we do want to
isolate only the TemplateIcon, we can address the viewbox instead.