Fix global search input unfocusing on new page
For unknown reasons, the search input would unfocus when it gets to
the search page. The fix for this was a random timeout to update
the search text. Not totally sure what was wrong or why this worked.

Also, fixed so you have to type 4 letters to start a search.