Fix foamcore selection dragging
Because I added a new element within the foamcore grid blanks, which is
an element as large as the blank, the selection wasn't working anymore
since selection requires an element be tagged as empty-space with a
`data-empty-space` html attribute. To fix this I added this attribute
to the new div I created so it'll be tagged as a empty space like
before.

I also did some renaming of things for consistency, I like `handle`
for events that are directly called from rendered elements.