Stop changing the prop data passed in
Before we were taking the array of breadcrumb items passed in and
directly changing them, which would lead them to be modified data.
This lead to inconsistencies as the base data from the component
above would have it's data modified by a subcomponent.

This changes that so everything uses copies of the data locally to
truncate and such, removing these problems.