Fix broken Shape charts x-axis
When Shape report items have less that 4 data points, the x-axis
is broken and has inconsistent behavior. This is due to us passing
a certain amount of tick count to the axis, forcing them to have more
ticks than they actually have which confuses VictoryAxis.

This change ensures the tick count is no more than the total amount
of values.