Various chart axis fixes for CD
Accounts for the point when a Creative Different chart has only one
data point. In this case it expands the domain 1 month before and after
the single data point so it renders the bar and label in the center. It
also expands the stroke width of the data so it' displays as a bar.

Also fixed a bug with the ovelapping data points due to a feature
we're not using yet. The feature allows CD to mark data points as
`prioritized` but it was causing other problems and is not in use
so I removed it for now.