Use num_columns, not STI type to check if a board
We eventually want all collections to act like foamcore boards. This is
an intermediary step, where we check if a collection has a num_columns
value to see if it is a board, not its single table inheritance type.

This commit makes changes for this on the frontend and backend.

This also ports some methods that previously only Collection::Board had
because they are needed in the serializer.