Setup caching and attributes
Will cache the results from the controller when creating a template
and in the activity builder. Will just send IDs in the serializer,
it's probably better to fetch all the collections when you open
the modal? Or async after components have mounted? It seems heavy
to include it all in the user / org payload.