Fix logic to set last used user templates
Before it wouldn't add to the array because the uniq logic was wrong
now it will move last used to top but always unique.

This changes a previous mistake where the template in GA tag manager
was actually the collection the template was being moved into, that
GA tag manager data should not correctly be the actual template.